refactor(workflow): move createLanguageModelToolResultOutput into workflow package
Its only consumer is WorkflowAgent. Export the lower-level primitives
(downloadAssets, createDefaultDownloadFunction, DownloadFunction) from
ai/internal and compose them in @ai-sdk/workflow instead.
