+++ packages/ai/internal/index.ts
@@ -8,10 +8,14 @@ export { convertAsyncIteratorToReadableStream } from '@ai-sdk/provider-utils';
 export { createAsyncIterableStream } from '../src/util/async-iterable-stream';
 export {
   convertToLanguageModelPrompt,
-  createLanguageModelToolResultOutput,
+  downloadAssets,
   mapToolResultOutput,
 } from '../src/prompt/convert-to-language-model-prompt';
 export { createToolModelOutput } from '../src/prompt/create-tool-model-output';
+export {
+  createDefaultDownloadFunction,
+  type DownloadFunction,
+} from '../src/util/download/download-function';
 export { prepareToolChoice } from '../src/prompt/prepare-tool-choice';
 export { prepareTools } from '../src/prompt/prepare-tools';
 export { standardizePrompt } from '../src/prompt/standardize-prompt';
