+++ packages/ai/src/prompt/convert-to-language-model-prompt.ts
@@ -17,7 +17,6 @@ import {
   type ReasoningFilePart,
   type ReasoningPart,
   type TextPart,
-  type Tool,
   type ToolCallPart,
   type ToolResultOutput,
   type ToolResultPart,
@@ -32,7 +31,6 @@ import type { Warning } from '../types/warning';
 import { InvalidMessageRoleError } from './invalid-message-role-error';
 import type { StandardizedPrompt } from './standardize-prompt';
 import { MissingToolResultsError } from '../error/missing-tool-result-error';
-import { createToolModelOutput } from './create-tool-model-output';
 
 export async function convertToLanguageModelPrompt({
   prompt,
@@ -428,7 +426,7 @@ function convertImagePartToFilePart(
 /**
  * Downloads files from URLs in the user messages.
  */
-async function downloadAssets(
+export async function downloadAssets(
   messages: ModelMessage[],
   download: DownloadFunction,
   supportedUrls: Record<string, RegExp[]>,
@@ -592,63 +590,6 @@ function convertPartToLanguageModelPart(
     data,
     providerOptions: part.providerOptions,
   };
-}
-
-export async function createLanguageModelToolResultOutput({
-  toolCallId,
-  toolName,
-  input,
-  output,
-  tool,
-  errorMode,
-  supportedUrls,
-  download = createDefaultDownloadFunction(),
-  provider,
-}: {
-  toolCallId: string;
-  toolName: string;
-  input: unknown;
-  output: unknown;
-  tool: Tool | undefined;
-  errorMode: 'none' | 'text' | 'json';
-  supportedUrls: Record<string, RegExp[]>;
-  download?: DownloadFunction;
-  provider?: string;
-}): Promise<LanguageModelV4ToolResultOutput> {
-  const modelOutput = await createToolModelOutput({
-    toolCallId,
-    input,
-    output,
-    tool,
-    errorMode,
-  });
-
-  const downloadedAssets =
-    modelOutput.type === 'content'
-      ? await downloadAssets(
-          [
-            {
-              role: 'tool',
-              content: [
-                {
-                  type: 'tool-result',
-                  toolCallId,
-                  toolName,
-                  output: modelOutput,
-                },
-              ],
-            },
-          ],
-          download,
-          supportedUrls,
-        )
-      : {};
-
-  return mapToolResultOutput({
-    output: modelOutput,
-    provider,
-    downloadedAssets,
-  });
 }
 
 export function mapToolResultOutput({
